FINALLY get stories to work with react router
